chore(mcp): extract insights visualization dispatch into pure modules
Verbatim move, no logic changes: inferVisualizationType and its shape guards move from Component.tsx to infer-visualization.ts, and toActorRows/ActorRow/InsightActorsData move from InsightActorsView.tsx to insightActorsTransforms.ts, so this decision logic can be unit-tested without pulling in React or chart dependencies.

Generated-By: PostHog Code
Task-Id: 86b4c62a-29c4-49d5-9e6c-c7b2b2c9953c
