+++ products/product_analytics/mcp/apps/InsightActorsView.tsx
@@ -3,41 +3,9 @@ import { type ReactElement, type ReactNode, useMemo } from 'react'
 import { DataTable, type DataTableColumn } from '@posthog/mcp-ui'
 import { Badge, Button } from '@posthog/quill'
 
-export interface InsightActorsData {
-    query: Record<string, unknown>
-    results: {
-        columns: string[]
-        results: (string | number | null | string[])[][]
-    }
-    hasMore: boolean
-    offset: number
-    _posthogUrl?: string
-}
+import { type ActorRow, type InsightActorsData, toActorRows } from './insightActorsTransforms'
 
-interface ActorRow {
-    distinct_id: string | null
-    email: string | null
-    name: string | null
-    event_count: number | null
-    recordings: string[]
-}
-
-function toActorRows(data: InsightActorsData): ActorRow[] {
-    const { columns, results } = data.results
-    return results.map((row) => {
-        const obj: Record<string, unknown> = {}
-        columns.forEach((col, i) => {
-            obj[col] = row[i]
-        })
-        return {
-            distinct_id: (obj.distinct_id as string) ?? null,
-            email: (obj.email as string) ?? null,
-            name: (obj.name as string) ?? null,
-            event_count: (obj.event_count as number) ?? null,
-            recordings: Array.isArray(obj.recordings) ? (obj.recordings as string[]) : [],
-        }
-    })
-}
+export type { InsightActorsData }
 
 interface InsightActorsViewProps {
     data: InsightActorsData
